Simplify setup function of system plugin
