+++ src/middlewared/middlewared/plugins/system/__init__.py
@@ -1,95 +1,11 @@
 # -*- coding=utf-8 -*-
-import asyncio
 import os
 import textwrap
 import uuid
 
 from middlewared.utils import run
 
-from .utils import FIRST_INSTALL_SENTINEL, lifecycle_conf, RE_KDUMP_CONFIGURED
-
-
-async def setup(middleware):
-    lifecycle_conf.SYSTEM_BOOT_ID = str(uuid.uuid4())
-
-    middleware.event_register('system', textwrap.dedent('''\
-        Sent on system state changes.
-
-        id=ready -- Finished boot process\n
-        id=reboot -- Started reboot process\n
-        id=shutdown -- Started shutdown process'''))
-
-    if os.path.exists('/tmp/.bootready'):
-        lifecycle_conf.SYSTEM_READY = True
-    else:
-        await firstboot(middleware)
-
-    settings = await middleware.call('system.general.config')
-    await middleware.call('core.environ_update', {'TZ': settings['timezone']})
-
-    middleware.logger.debug(f'Timezone set to {settings["timezone"]}')
-
-    await middleware.call('system.general.set_language')
-    await middleware.call('system.general.set_crash_reporting')
-
-    middleware.event_subscribe('system', _event_system)
-    CRASH_DIR = '/data/crash'
-    os.makedirs(CRASH_DIR, exist_ok=True)
-    os.chmod(CRASH_DIR, 0o775)
-
-    await middleware.call('sysctl.set_zvol_volmode', 2)
-
-    middleware.register_hook('system.post_license_update', hook_license_update)
-
-
-async def hook_license_update(middleware, prev_product_type, *args, **kwargs):
-    if prev_product_type != 'ENTERPRISE' and await middleware.call('system.product_type') == 'ENTERPRISE':
-        await middleware.call('system.advanced.update', {'autotune': True})
-
-
-async def _event_system(middleware, event_type, args):
-
-    if args['id'] == 'ready':
-        lifecycle_conf.SYSTEM_READY = True
-
-        # Check if birthday is already set
-        birthday = await middleware.call('system.birthday')
-        if birthday is None:
-            # try to set birthday in background
-            asyncio.ensure_future(_update_birthday(middleware))
-
-        if (await middleware.call('system.advanced.config'))['kdump_enabled']:
-            cp = await run(['kdump-config', 'status'], check=False)
-            if cp.returncode:
-                middleware.logger.error('Failed to retrieve kdump-config status: %s', cp.stderr.decode())
-            else:
-                if not RE_KDUMP_CONFIGURED.findall(cp.stdout.decode()):
-                    await middleware.call('alert.oneshot_create', 'KdumpNotReady', None)
-                else:
-                    await middleware.call('alert.oneshot_delete', 'KdumpNotReady', None)
-        else:
-            await middleware.call('alert.oneshot_delete', 'KdumpNotReady', None)
-
-        if await middleware.call('system.first_boot'):
-            asyncio.ensure_future(middleware.call('usage.firstboot'))
-
-    if args['id'] == 'shutdown':
-        lifecycle_conf.SYSTEM_SHUTTING_DOWN = True
-
-
-async def _update_birthday(middleware):
-    while True:
-        birthday = await middleware.call('system.get_synced_clock_time')
-        if birthday:
-            middleware.logger.debug('Updating birthday data')
-            # update db with new birthday
-            settings = await middleware.call('datastore.config', 'system.settings')
-            await middleware.call(
-                'datastore.update', 'system.settings', settings['id'], {'stg_birthday': birthday}, {'ha_sync': False}
-            )
-            break
-        else:
-            await asyncio.sleep(300)
+from .utils import FIRST_INSTALL_SENTINEL, lifecycle_conf
 
 
 async def firstboot(middleware):
@@ -118,3 +34,33 @@ async def firstboot(middleware):
                 middleware.logger.error(
                     'Failed to set keep attribute for Initial-Install boot environment: %s', cp.stderr.decode()
                 )
+
+
+async def setup(middleware):
+    lifecycle_conf.SYSTEM_BOOT_ID = str(uuid.uuid4())
+
+    middleware.event_register('system', textwrap.dedent('''\
+        Sent on system state changes.
+
+        id=ready -- Finished boot process\n
+        id=reboot -- Started reboot process\n
+        id=shutdown -- Started shutdown process'''))
+
+    if os.path.exists('/tmp/.bootready'):
+        lifecycle_conf.SYSTEM_READY = True
+    else:
+        await firstboot(middleware)
+
+    settings = await middleware.call('system.general.config')
+    await middleware.call('core.environ_update', {'TZ': settings['timezone']})
+
+    middleware.logger.debug(f'Timezone set to {settings["timezone"]}')
+
+    await middleware.call('system.general.set_language')
+    await middleware.call('system.general.set_crash_reporting')
+
+    CRASH_DIR = '/data/crash'
+    os.makedirs(CRASH_DIR, exist_ok=True)
+    os.chmod(CRASH_DIR, 0o775)
+
+    await middleware.call('sysctl.set_zvol_volmode', 2)
